Use DefaultAWSCredentialProviderChain over hardcoded credentials
Allow for using the DefaultAWSCredentialProviderChain in place
of requiring hardcoded access keys in the flume config.  This
will allow using EC2 Instance Profiles and other methods
to access Kinesis so users can avoid putting sensitive keys
in plain text.

The existing hardcoded credentials are still supported for
now to allow for backward compatability.  If those are present
in the config they take precedence over the Default chain.